pyproject.toml
[tool.poetry]
name = "plate-detection"
version = "0.1.0"
description = ""
authors = ["Edgar454 <mevaed4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
transformers = "4.40.0"
pillow = "^10.4.0"
torch = "^2.4.0"
tqdm = "^4.66.5"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
